Record non-API request failures and cover RequestDiagnostics with tests
