+++ Ruddarr/Utilities/RequestDiagnostics.swift
@@ -6,9 +6,7 @@ actor RequestDiagnostics {
     private var entries: [FailedRequest] = []
     private let limit = 50
 
-    func record(method: String, url: String, instance: String?, error: API.Error) {
-        let reason = FailedRequest.Reason(error)
-
+    func record(method: String, url: String, instance: String?, reason: FailedRequest.Reason) {
         if let last = entries.first, last.method == method, last.url == url, last.reason == reason {
             return
         }
@@ -70,39 +68,6 @@ struct FailedRequest: Identifiable, Equatable, Sendable {
         case .status(_, let message): message
         case .decoding(let text): text
         case .transport(let text): text
-        }
-    }
-}
-
-extension FailedRequest.Reason {
-    init(_ error: API.Error) {
-        switch error {
-        case .badStatusCode(let code):
-            self = .status(code: code, message: nil)
-        case .errorResponse(let code, let message):
-            self = .status(code: code, message: message)
-        case .decodingError(let error):
-            let path = error.context.codingPath.map(\.stringValue).joined(separator: ".")
-            let description = error.context.debugDescription
-            self = .decoding(path.isEmpty ? description : "\(path): \(description)")
-        case .notConnectedToInternet:
-            self = .transport("Not connected to the internet")
-        case .timeoutOnPrivateIp(let error):
-            self = .transport(error.localizedDescription)
-        case .urlError(let error):
-            self = .transport(error.localizedDescription)
-        case .nsError(let error):
-            self = .transport(error.localizedDescription)
-        case .localizedError(let error):
-            self = .transport(error.errorDescription ?? String(describing: error))
-        case .appError(let error):
-            self = .transport(error.errorDescription ?? String(describing: error))
-        case .invalidUrl(let url):
-            self = .transport("Invalid URL: \(url)")
-        case .error(let error):
-            self = .transport(String(describing: error))
-        case .void:
-            self = .transport("Request failed")
         }
     }
 }
